[#12]Rename: SecurityConfig to WebSecurityConfig
