--- src/main/java/com/techeer/port/voilio/global/config/security/SecurityConfig.java
@@ -1,48 +1,0 @@
-package com.techeer.port.voilio.global.config.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-  private final JwtProvider jwtProvider;
-
-  public SecurityConfig(JwtProvider jwtProvider) {
-    this.jwtProvider = jwtProvider;
-  }
-
-  @Bean
-  public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-    http.csrf()
-        .disable()
-        .sessionManagement()
-        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-        .and()
-        .formLogin()
-        .disable()
-        .httpBasic()
-        .disable()
-        .authorizeRequests()
-        .antMatchers("/api/v1/users")
-        .permitAll()
-        .and()
-        .addFilterBefore( // JwtAuthenticationFilter 를 UsernamePasswordAuthenticationFilter 전에 넣는다
-            new JwtAuthenticationFilter(jwtProvider), UsernamePasswordAuthenticationFilter.class);
-
-    return http.build();
-  }
-
-  @Bean
-  public PasswordEncoder passwordEncoder() {
-    return new BCryptPasswordEncoder();
-  }
-}
